Move analysis-related code.
