+++ assets/fav_bright.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6d69e1c0ac8961625492ebb5d322ab074e5dccfc983b284f2f6147fa27e2b3d0
+size 24776
