+++ assets/fav_dark.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:14099217558eea074f09343c9fdda17f5d14c6f262fd82cbbda7e39f0489533b
+size 24499
