+++ assets/logo_bright.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ddc33229ad50b93b52b8b7e387ae32abd55415689f0f0a8448fbd61a5955da85
+size 84982
