+++ assets/logo_dark.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:4e2ce8c2a409d97db8ee55b038603e91d55ac5520207191c8fc72c891d4e7189
+size 73442
