+++ columnflow/calibration/example.py
@@ -1,11 +1,10 @@
 # coding: utf-8
 
 """
-Calibration methods for testing purposes.
+Exemplary calibration methods.
 """
 
 from columnflow.calibration import Calibrator, calibrator
-from columnflow.production.seeds import deterministic_seeds
 from columnflow.util import maybe_import
 from columnflow.columnar_util import set_ak_column
 
@@ -23,7 +22,7 @@ ak = maybe_import("awkward")
         "Jet.pt_jec_down", "Jet.mass_jec_down",
     },
 )
-def jec_test(self: Calibrator, events: ak.Array, **kwargs) -> ak.Array:
+def example(self: Calibrator, events: ak.Array, **kwargs) -> ak.Array:
     # a) "correct" Jet.pt by scaling four momenta by 1.1 (pt<30) or 0.9 (pt<=30)
     # b) add 4 new columns representing the effect of JEC variations
 
@@ -41,17 +40,5 @@ def jec_test(self: Calibrator, events: ak.Array, **kwargs) -> ak.Array:
     set_ak_column(events, "Jet.mass_jec_up", events.Jet.mass * 1.05)
     set_ak_column(events, "Jet.pt_jec_down", events.Jet.pt * 0.95)
     set_ak_column(events, "Jet.mass_jec_down", events.Jet.mass * 0.95)
-
-    return events
-
-
-@calibrator(
-    uses={jec_test, deterministic_seeds},
-    produces={jec_test, deterministic_seeds},
-)
-def test(self: Calibrator, events: ak.Array, **kwargs) -> ak.Array:
-    self[jec_test](events, **kwargs)
-
-    self[deterministic_seeds](events, **kwargs)
 
     return events
