--- columnflow/scripts/hello_world.py
@@ -1,14 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-"""
-Script that prints "Hello world!".
-"""
-
-
-def main():
-    print("Hello world!")
-
-
-if __name__ == "__main__":
-    main()
