+++ law.cfg
@@ -1,8 +1,3 @@
-[core]
-
-extend_configs: law_user.cfg
-
-
 [modules]
 
 columnflow.tasks.framework
@@ -20,26 +15,30 @@ columnflow.tasks.cutflow
 
 [analysis]
 
-default_analysis: analysis_st
+default_analysis: columnflow.example_config.analysis_st.analysis_st
 default_config: run2_pp_2018
 default_dataset: st_tchannel_t
 
-production_modules: columnflow.production.{categories,processes,features,pileup,normalization,seeds,weights}
-calibration_modules: columnflow.calibration.{test,jets}
-selection_modules: columnflow.selection.test
-ml_modules: columnflow.ml.test
-inference_modules: columnflow.inference.test
+production_modules: columnflow.production.{example,categories,processes,pileup,normalization,seeds}
+calibration_modules: columnflow.calibration.{example,jets}
+selection_modules: columnflow.selection.example
+ml_modules: columnflow.ml.example
+inference_modules: columnflow.inference.example
 
 
 [outputs]
 
 # list of all used file systems
-wlcg_file_systems: wlcg_fs, wlcg_fs_desy, wlcg_fs_cernbox, wlcg_fs_software, wlcg_fs_desy_store, wlcg_fs_infn_redirector
+wlcg_file_systems: wlcg_fs, wlcg_fs_infn_redirector, wlcg_fs_global_redirector
+
+# list of file systems used by columnflow.tasks.external.GetDatasetLFNs.iter_nano_files to
+# look for the correct fs per nano input file (in that order)
+wlcg_lfn_sources: wlcg_fs_infn_redirector, wlcg_fs_global_redirector
 
 # output locations per task family
 # for local targets : "local[, STORE_PATH]"
 # for remote targets: "wlcg[, WLCG_FS_NAME]"
-; CalibrateEvents: wlcg, wlcg_fs_cernbox
+; CalibrateEvents: wlcg, wlcg_fs_...
 
 
 [job]
@@ -57,42 +56,12 @@ gfal2: WARNING
 
 [wlcg_fs]
 
-base: &::wlcg_fs_desy::base
+base: gsiftp://dcache-door-cms04.desy.de:2811/pnfs/desy.de/cms/tier2/store/user/$CF_CERN_USER/$CF_STORE_NAME
 create_file_dir: True
 use_cache: $CF_WLCG_USE_CACHE
 cache_root: $CF_WLCG_CACHE_ROOT
 cache_cleanup: $CF_WLCG_CACHE_CLEANUP
 cache_max_size: 50GB
-
-
-[wlcg_fs_desy]
-
-xrootd_base: root://dcache-cms-xrootd.desy.de:1094/pnfs/desy.de/cms/tier2/store/user/$CF_CERN_USER/$CF_STORE_NAME
-gsiftp_base: gsiftp://dcache-door-cms04.desy.de:2811/pnfs/desy.de/cms/tier2/store/user/$CF_CERN_USER/$CF_STORE_NAME
-base: &::gsiftp_base
-
-
-[wlcg_fs_cernbox]
-
-xrootd_base: root://eosuser.cern.ch/eos/user/$CF_CERN_USER_FIRSTCHAR/$CF_CERN_USER/$CF_STORE_NAME
-base: &::xrootd_base
-
-
-[wlcg_fs_software]
-
-base: &::wlcg_fs_desy::base
-
-
-[wlcg_fs_desy_store]
-
-xrootd_base: root://dcache-cms-xrootd.desy.de:1094/pnfs/desy.de/cms/tier2
-gsiftp_base: gsiftp://dcache-door-cms04.desy.de:2811/pnfs/desy.de/cms/tier2
-base: &::xrootd_base
-use_cache: $CF_WLCG_USE_CACHE
-cache_root: $CF_WLCG_CACHE_ROOT
-cache_cleanup: $CF_WLCG_CACHE_CLEANUP
-cache_max_size: 15GB
-cache_global_lock: True
 
 
 [wlcg_fs_infn_redirector]
