+++ sandboxes/cf_dev.sh
@@ -6,11 +6,10 @@
 action() {
     local this_file="$( [ ! -z "$ZSH_VERSION" ] && echo "${(%):-%x}" || echo "${BASH_SOURCE[0]}" )"
     local this_dir="$( cd "$( dirname "$this_file" )" && pwd )"
-    local cf_dir="$( dirname "$this_dir" )"
 
     # set variables and source the generic venv setup
     export CF_VENV_NAME="$( basename "${this_file%.sh}" )"
-    export CF_VENV_REQUIREMENTS="$cf_dir/requirements_dev.txt,$cf_dir/requirements_prod.txt"
+    export CF_VENV_REQUIREMENTS="$CF_BASE/requirements_dev.txt,$CF_BASE/requirements_prod.txt"
 
     source "$this_dir/_setup_venv.sh" "$@"
 }
